+++ packages/frameworks/react/src/number-input/number-input.test.tsx
@@ -52,7 +52,7 @@ describe('NumberInput', () => {
     })
   })
 
-  it('should  allow value to exceed max when allowOverflow is true', async () => {
+  it('should allow value to exceed max when allowOverflow is true', async () => {
     render(<ComponentUnderTest allowOverflow max={10} defaultValue="15" />)
     const input = screen.getByRole('spinbutton')
     expect(input).toHaveValue('15')