+++ packages/frameworks/solid/src/radio-group/radio-group.test.tsx
@@ -3,30 +3,31 @@ import { render, screen } from '@solidjs/testing-library'
 import user from '@testing-library/user-event'
 import { For } from 'solid-js'
 import { vi } from 'vitest'
-import { getParts } from '../setup-test'
+import { getExports, getParts } from '../setup-test'
 import { RadioGroup, type RadioGroupProps } from './'
 
-const options = [
-  { id: 'apple', label: 'Apples' },
-  { id: 'orange', label: 'Oranges' },
-  { id: 'mango', label: 'Mangoes', disabled: true },
-  { id: 'grape', label: 'Grapes' },
-]
-
-const ComponentUnderTest = (props: RadioGroupProps) => (
-  <RadioGroup.Root {...props}>
-    <RadioGroup.Label>Fruits</RadioGroup.Label>
-    <RadioGroup.Indicator />
-    <For each={options}>
-      {(option) => (
-        <RadioGroup.Item value={option.id} disabled={option.disabled}>
-          <RadioGroup.ItemText>{option.label}</RadioGroup.ItemText>
-          <RadioGroup.ItemControl />
-        </RadioGroup.Item>
-      )}
-    </For>
-  </RadioGroup.Root>
-)
+const ComponentUnderTest = (props: RadioGroupProps) => {
+  const items = [
+    { label: 'React', value: 'react' },
+    { label: 'Solid', value: 'solid' },
+    { label: 'Vue', value: 'vue' },
+    { label: 'Svelte', value: 'svelte', disabled: true },
+  ]
+  return (
+    <RadioGroup.Root {...props}>
+      <RadioGroup.Label>Fruits</RadioGroup.Label>
+      <RadioGroup.Indicator />
+      <For each={items}>
+        {(item) => (
+          <RadioGroup.Item value={item.value} disabled={item.disabled}>
+            <RadioGroup.ItemText>{item.label}</RadioGroup.ItemText>
+            <RadioGroup.ItemControl />
+          </RadioGroup.Item>
+        )}
+      </For>
+    </RadioGroup.Root>
+  )
+}
 
 describe('Radio Group', () => {
   it.each(getParts(radioGroupAnatomy))('should render part! %s', async (part) => {
@@ -34,12 +35,17 @@ describe('Radio Group', () => {
     expect(document.querySelector(part)).toBeInTheDocument()
   })
 
+  it.each(getExports(radioGroupAnatomy))('should export %s', async (part) => {
+    expect(RadioGroup[part]).toBeDefined()
+  })
+
   it('should invoke onValueChange if another value has selected', async () => {
     const onValueChange = vi.fn()
+
     render(() => <ComponentUnderTest onValueChange={onValueChange} />)
 
-    await user.click(screen.getByLabelText('Grapes'))
-    expect(onValueChange).toHaveBeenCalledWith({ value: 'grape' })
+    await user.click(screen.getByLabelText('Solid'))
+    expect(onValueChange).toHaveBeenCalledWith({ value: 'solid' })
   })
 
   it('should not invoke onValueChange if option is disabled', async () => {
@@ -47,7 +53,7 @@ describe('Radio Group', () => {
 
     render(() => <ComponentUnderTest onValueChange={onValueChange} />)
 
-    await user.click(screen.getByLabelText('Mangoes'))
+    await user.click(screen.getByLabelText('Svelte'))
     expect(onValueChange).not.toHaveBeenCalled()
   })
 })