+++ packages/frameworks/solid/src/splitter/splitter.test.tsx
@@ -1,4 +1,6 @@
+import { splitterAnatomy } from '@ark-ui/anatomy'
 import { render } from '@solidjs/testing-library'
+import { getExports, getParts } from '../setup-test'
 import { Splitter, type SplitterProps } from './'
 
 const ComponentUnderTest = (props: SplitterProps) => (
@@ -18,7 +20,12 @@ const ComponentUnderTest = (props: SplitterProps) => (
 )
 
 describe('Splitter', () => {
-  it('should render', async () => {
+  it.each(getParts(splitterAnatomy))('should render part! %s', async (part) => {
     render(() => <ComponentUnderTest />)
+    expect(document.querySelector(part)).toBeInTheDocument()
+  })
+
+  it.each(getExports(splitterAnatomy))('should export %s', async (part) => {
+    expect(Splitter[part]).toBeDefined()
   })
 })