+++ packages/frameworks/solid/CHANGELOG.md
@@ -10,6 +10,10 @@ description: All notable changes to this project will be documented in this file
 
 - Added `FileUpload` component
 
+### Fixed
+
+- Resolved an issue where `Dialog.CloseTrigger` had `data-part` as `trigger` instead of `close-trigger`.
+
 ## [1.0.1] - 2023-11-10
 
 ### Fixed